% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{bool_extractor}
\alias{bool_extractor}
\title{Format observation read from .yxdb Connection.}
\usage{
bool_extractor(start, field_len, buffer)
}
\description{
Internal functions to convert raw data to various data types.
}
