% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_header_and_field_info.R
\name{get_header_and_field_info}
\alias{get_header_and_field_info}
\title{Extract Header and Meta-Information from .yxdb Connection}
\usage{
get_header_and_field_info(path)
}
\arguments{
\item{path}{Path to .yxdb file.}
}
\value{
A list containing file meta-information and initialized parameters for file reading.
}
\description{
Reads header and collects meta-information from a .yxdb connection into a list.
}
\details{
.yxdb files reserve their first bits for information about the file.
This includes the number of observations, field names, and field types.
`get_header_and_field_info` parses the first bits of a .yxdb connection
and returns a list containing field information, functions for extracting
data from fields of different formats, and initialized parameters for
reading and (most likely) decompressing the .yxdb file.
}
\examples{
get_header_and_field_info()
}
