% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/record_reader.R
\name{next_record}
\alias{next_record}
\title{Collect next observation from .yxdb Connection.}
\usage{
next_record(yxdb)
}
\arguments{
\item{yxdb}{.yxdb file connection as initialized by `get_header_and_field_info`.}
}
\value{
yxdb overwritten and staged for next observation.
}
\description{
Checks if next observation is last and, if not, calls for next record to be read.
}
\examples{
next_record()
}
