% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bits.R
\docType{data}
\name{Ox80000000}
\alias{Ox80000000}
\title{Unsigned 32-bit integers.}
\format{
An object of class \code{raw} of length 32.
}
\usage{
Ox80000000
}
\description{
Storing frequently used, and sometimes difficult to encode, 32-bit integers.
}
\keyword{datasets}
