% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractors.R
\name{bool_extractor}
\alias{bool_extractor}
\title{Format observation read from .yxdb Connection.}
\usage{
bool_extractor(start, field_len, buffer)
}
\description{
Internal functions to convert raw data to various data types.
}
\details{
R has native support for bool, byte, int32, double, character, and date only.
However, .yxdb files may contain a variety of additional field types. These
extractor functions are a brute force solution to coerce non-native data
types to something usable by R. In some cases (e.g., int64), data conversion
will occur by necessity.
}
